refactor(api): reorganize include layout and restore banner
Refs #90
