+++ include/framekit/service/bootstrap.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "framekit/runtime/service_context.hpp"
+#include "framekit/service/context.hpp"
 
 #include <chrono>
 #include <cstdint>
