+++ include/framekit/spec/process_role.hpp
