+++ src/event_bus.cpp
@@ -1,4 +1,4 @@
-#include "framekit/runtime/event_bus.hpp"
+#include "framekit/event/bus.hpp"
 
 #include <algorithm>
 #include <stdexcept>
